feat: working payments and trialing
